initial refactoring. separated Classes out into individual rb files.
