+++ README.md
@@ -5,7 +5,7 @@ Core Features
 [x] Add a todo
 [x] Edit a todo
 [x] Mark a todo as complete
-- Filter todos by "All", "Active" and "Complete"
+[x] Filter todos by "All", "Active" and "Complete"
 - It needs to look good ;)
 
 Stretch goals
