+++ todo_item.rb
@@ -1,0 +1,17 @@
+require 'sinatra'
+
+class ToDoItem
+  def initialize(todo)
+    @todo = todo
+    @status = "active"
+    @filter = "all"
+  end
+
+  def todo_list_entry
+    {
+      :action => @todo,
+      :status => @status,
+      :filter => @filter
+    }
+  end
+end
