+++ views/index.erb
@@ -1,8 +1,10 @@
 <html>
-  <head></head>
-  <body>
+  <head>
 
-    <h1><strong>WELCOME TO YOUR TO DO LIST</strong></h1>
+  </head>
+
+  <body>
+    <h2><strong>WELCOME TO YOUR TO DO LIST</strong></h2>
 
     <form>
       <input type='text' name='add_item' />
@@ -10,20 +12,25 @@
     </form>
 
     <form>
-    <input type='text' name='find' value='name of action' />
-    <input type='text' name='edit' value='change to'  /><input type='submit' value="Edit an action"/>
+      <input type='text' name='find' value='name of action' />
+      <input type='text' name='edit' value='change to'  />
+      <input type='submit' value="Edit an action"/>
     </form>
 
     <form>
-    <input type='text' name='complete' value='mark as complete'/><input type='submit' value="Complete an action"/>
+      <input type='text' name='complete' value='mark as complete'/>
+      <input type='submit' value="Complete an action"/>
     </form>
 
     <form>
-      <p>Filter list by: <input type='submit' name="display_all" value="All"/><input type='submit' name="display_active" value="Active"/><input type='submit' name="display_complete" value="Completed"/>
-        </p>
-      </form>
+      <p>
+        Filter list by: <input type='submit' name="display_all" value="All"/><input type='submit' name="display_active" value="Active"/><input type='submit' name="display_complete" value="Completed"/>
+      </p>
+    </form>
 
-    <h4><strong>Here is your ToDo List:</strong></h4><br> <%= LIST_OF_ACTIONS %> <%= display_list %>
+    <h4><strong>Here is your filtered ToDo List:</strong></h4>
+    <p><%= display_list %></p>
 
   </body>
+
 </html>
